Organises the proxy objects into anonymous_review.py
